requirements.txt
asttokens==2.0.8
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==22.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.5
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==5.1.1
distro==1.7.0
distro-info===1.1build1
dnspython==2.2.1
email-validator==1.3.0
executing==1.0.0
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.37.3
forex-python==1.8
greenlet==1.1.3.post0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.23.3
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
psycopg2-pool==1.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
Pygments==2.13.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-xlib==0.31
PythonTurtle==0.3.2
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
SecretStorage==3.3.1
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
sos==4.3
SQLAlchemy==1.4.42
ssh-import-id==5.11
stack-data==0.5.0
systemd-python==234
traitlets==5.4.0
Twisted==22.1.0
ubuntu-advantage-tools==27.9
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
XWindows==0.0.7
zipp==1.0.0
zope.interface==5.4.0
